backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==4.0.6
django-cors-headers==4.3.1
django-redis==5.2.0
django-resized==1.0.3
django-stubs==5.1.1
django-stubs-ext==5.1.1
djangorestframework==3.14.0
djangorestframework-api-key==3.0.0
djangorestframework-simplejwt==5.3.1
flake8==7.1.1
gunicorn==20.0.4
idna==3.10
kombu==5.4.2
mccabe==0.7.0
mypy==1.13.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.6
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
random-username==1.0.2
redis==5.0.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.6.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
types-PyYAML==6.0.12.20241221
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
pillow==11.0.0
black==24.10.0
django-mptt==0.16.0
django-mdeditor==0.1.20
django-select2==8.2.3
Markdown==3.7
markdownify==0.14.1
